requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
docker-pycreds==0.4.0
gitdb==4.0.10
GitPython==3.1.29
hydra-core==1.2.0
idna==3.4
numpy==1.23.5
omegaconf==2.2.3
packaging==21.3
pathtools==0.1.2
Pillow==9.3.0
promise==2.3
protobuf==4.21.10
psutil==5.9.4
pycparser==2.21
pyparsing==3.0.9
PySoundFile==0.9.0.post1
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
sentry-sdk==1.11.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
torch==1.13.0
torchaudio==0.13.0
torchsummary==1.5.1
torchvision==0.14.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
wandb==0.13.5
